Added note for developers
I am interested in contributing to the project, so first I need to understand the source code. However the few comments in the code itself didn't help much, so when viewing the *docs* to see if there was something useful I found an overview of the library code which really helped me. It might also help other future contributors.